New changes for mediator pipelines for AddUserCommandHandler
