+++ myn-graphql-sample/Data/Handlers/Commands/AddUserCommandHandler.cs
@@ -1,41 +1,9 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using myn_graphql_sample.Data.Requests.Commands;
 using myn_graphql_sample.Entities;
-using myn_graphql_sample.Data.Context;
 
 namespace myn_graphql_sample.Data.Handlers.Commands
 {
-    public class TransactionBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-    {
-        private readonly AppDbContext _context;
-
-        public TransactionBehavior(AppDbContext context)
-        {
-            _context = context;
-        }
-
-        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
-        {
-            using (var transaction = await _context.Database.BeginTransactionAsync())
-            {
-                try
-                {
-                    var response = await next();
-                    await _context.SaveChangesAsync();
-                    transaction.Commit();
-                    return response;
-                }
-                catch
-                {
-                    transaction.Rollback();
-                    throw;
-                }
-            }
-        }
-    }
-
     public class AddUserCommandHandler : IRequestHandler<AddUserCommand, User>
     {
         private readonly AppDbContext _context;
@@ -47,8 +15,11 @@ namespace myn_graphql_sample.Data.Handlers.Commands
 
         public async Task<User> Handle(AddUserCommand request, CancellationToken cancellationToken)
         {
-            _context.Users.Add(request.Input);
-            return request.Input;
+            _context.Users.Add(request.input);
+            _context.SaveChanges();
+
+            return request.input;
         }
+
     }
 }
